refactor: satisfy child execution lint gates
