+++ pkg/services/factory_runtime/internal/services/orchestration/javascript/runtime/host_primitives_test.go
@@ -6,8 +6,53 @@ import (
 	"testing"
 
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime"
+	workflowruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/javascript/runtime"
 	workflowpolicy "github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/orchestratorcontract"
 )
+
+func TestResolveChildWorkerSettings_AntigravityUsesModelEmbeddedEffort(t *testing.T) {
+	for _, executorProvider := range []string{"", "SCRIPT_WRAP"} {
+		t.Run("executorProvider="+executorProvider, func(t *testing.T) {
+			got, err := workflowruntime.ResolveChildWorkerSettings(
+				factory.JavaScriptChildExecutionRequest{
+					ExecutorProvider: executorProvider,
+					ModelProvider:    " ANTIGRAVITY ",
+					Model:            "gemini-3.6-flash-medium",
+					ReasoningEffort:  " HIGH ",
+				},
+				nil,
+				factory.JavaScriptWorkerSettings{},
+			)
+			if err != nil {
+				t.Fatalf("ResolveChildWorkerSettings() error = %v", err)
+			}
+			if got.ModelProvider != "antigravity" || got.Model != "gemini-3.6-flash-medium" {
+				t.Fatalf("selection = %#v, want canonical Antigravity provider and exact model", got)
+			}
+			if got.ReasoningEffort != "" {
+				t.Fatalf("selection reasoning effort = %q, want omitted for model-encoded Antigravity effort", got.ReasoningEffort)
+			}
+		})
+	}
+}
+
+func TestResolveChildWorkerSettings_CodexRetainsReasoningEffort(t *testing.T) {
+	got, err := workflowruntime.ResolveChildWorkerSettings(
+		factory.JavaScriptChildExecutionRequest{
+			ModelProvider:   "CODEX",
+			Model:           "gpt-5.6-luna",
+			ReasoningEffort: " XHIGH ",
+		},
+		nil,
+		factory.JavaScriptWorkerSettings{},
+	)
+	if err != nil {
+		t.Fatalf("ResolveChildWorkerSettings() error = %v", err)
+	}
+	if got.ModelProvider != "codex" || got.Model != "gpt-5.6-luna" || got.ReasoningEffort != "xhigh" {
+		t.Fatalf("selection = %#v, want codex/gpt-5.6-luna/xhigh", got)
+	}
+}
 
 func TestRun_ProgressPrimitives_EmitsOrderedRuntimeRecords(t *testing.T) {
 	source := readFixture(t, "progress-primitives.workflow.js")
